requirements.txt
# EMG-PKRI 项目依赖（Docker 友好版本）

# 核心依赖
pyyaml>=6.0              # 读取 config.yaml 配置文件

# LLM API
google-genai>=0.1.0      # Gemini API SDK（当前使用 gemma-3-27b-it）

# 数据处理（Day1 未使用，但后续可能用到）
pandas>=2.0.0            # 数据处理和分析
numpy>=1.24.0            # 数值计算
scikit-learn>=1.3.0      # 机器学习（后续模型训练可能用到）

# 测试
pytest>=7.4.0            # 单元测试框架
pytest-cov>=4.1.0        # 测试覆盖率

# 模型训练（Day2）
torch>=2.0.0             # PyTorch
transformers>=4.35.0     # HuggingFace Transformers
peft>=0.6.0              # Parameter-Efficient Fine-Tuning (LoRA)
accelerate>=0.24.0       # 训练加速
bitsandbytes>=0.41.0     # 量化支持（可选，用于 8-bit 训练）

# 注意：
# - 所有标准库（os, json, time, logging, signal, sys, re, hashlib, argparse, pathlib, typing, abc, dataclasses）不需要安装
# - API Key 通过环境变量 GEMINI_API_KEY 提供，不需要 python-dotenv


